merch_service/requirements.txt
asgiref==3.8.1
black==25.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
coverage==7.6.12
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.17.0
flake8==7.1.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
gunicorn==23.0.0
identify==2.6.7
idna==3.10
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.32.9
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.1.0
psutil==6.1.1
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.2
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
